# packages/types/src/openapi.yaml
openapi: 3.0.3
info:
  title: UR10 Robot Kiosk API
  description: |
    REST API for controlling UR10 robot and managing chess games in kiosk mode.
    
    This API provides endpoints for:
    - Robot connection and control
    - Real-time telemetry via WebSocket
    - Chess game management with engine integration
    - Session management and authentication
    - System configuration and settings
    
    ## Authentication
    
    The API uses PIN-based authentication with different access levels:
    - **Operator**: Basic robot control and chess gameplay
    - **Supervisor**: Advanced settings and safety overrides  
    - **Admin**: Full system configuration and maintenance
    
    ## WebSocket Endpoints
    
    Real-time communication is handled via WebSocket connections:
    - `/ws/telemetry` - Robot telemetry and status updates
    - `/ws/chess` - Chess game state updates
    - `/ws/system` - System alerts and notifications
    
  version: 1.0.0
  contact:
    name: UR10 Kiosk Team
    email: support@ur10-kiosk.local
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://localhost:8000
    description: Local development server
  - url: https://ur10-kiosk.local:8000
    description: Production kiosk server
...
paths:
...
  /api/robot/connect:
    post:
      tags: [Robot Control]
      summary: Connect to robot
      description: Establish connection to UR10 robot
      security:
        - SessionAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                hostname:
                  type: string
                  description: Robot IP address or hostname
                  example: "192.168.1.100"
                port:
                  type: integer
                  description: Robot port number
                  example: 30004
                timeout:
                  type: integer
                  description: Connection timeout in milliseconds
                  example: 5000
      responses:
        '200':
          description: Connection successful
        '400':
          description: Invalid connection parameters
        '503':
          description: Unable to connect to robot
  /api/robot/disconnect:
    post:
      tags: [Robot Control]
      summary: Disconnect from robot
      description: Close connection to UR10 robot
      security:
        - SessionAuth: []
      responses:
        '200':
          description: Disconnection successful
        '400':
          description: Robot not connected
  /api/robot/home:
    post:
      tags: [Robot Control]
      summary: Home robot
      description: Move robot to home position
      security:
        - SessionAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                speed:
                  type: number
                  description: Movement speed (m/s)
                  example: 0.1
                acceleration:
                  type: number
                  description: Movement acceleration (m/s²)
                  example: 0.5
      responses:
        '200':
          description: Homing started
        '400':
          description: Robot not connected or invalid parameters
        '409':
          description: Robot busy or in error state
  /api/robot/stop:
    post:
      tags: [Robot Control]
      summary: Stop robot
      description: Stop current robot movement
      security:
        - SessionAuth: []
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                emergency:
                  type: boolean
                  description: Emergency stop (true) or normal stop (false)
                  example: false
      responses:
        '200':
          description: Stop command sent
        '400':
          description: Robot not connected
  /api/robot/jog:
    post:
      tags: [Robot Control]
      summary: Jog robot
      description: Move robot incrementally in specified direction
      security:
        - SessionAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/JogRequest'
      responses:
        '200':
          description: Jog movement started
        '400':
          description: Invalid jog parameters
        '409':
          description: Robot busy or in error state
...
  /api/system/logs:
    get:
      tags: [System]
      summary: Get system logs
      description: Retrieve system logs
      security:
        - SessionAuth: [supervisor]
      parameters:
        - name: level
          in: query
          description: Log level filter
          schema:
            type: string
            enum: [debug, info, warn, error]
        - name: limit
          in: query
          description: Maximum number of log entries
          schema:
            type: integer
            default: 100
        - name: start
          in: query
          description: Start timestamp
          schema:
            type: integer
      responses:
        '200':
          description: Log entries
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    timestamp:
                      type: integer
                    level:
                      type: string
                    message:
                      type: string
                    component:
                      type: string
                    details:
                      type: object
components:
  securitySchemes:
    SessionAuth:
      type: apiKey
      in: header
      name: X-Session-ID
      description: Session ID obtained from login
  schemas:
...
    JogRequest:
      type: object
      required: [mode, axis, direction, distance, speed]
      properties:
        mode:
          type: string
          enum: [tcp, joint]
        axis:
          type: string
          enum: [x, y, z, rx, ry, rz, base, shoulder, elbow, wrist1, wrist2, wrist3]
        direction:
          type: string
          enum: [positive, negative]
        distance:
          type: number
          minimum: 0
        speed:
          type: number
          minimum: 0
        acceleration:
          type: number
          minimum: 0
...